feat: Separating App Initialization from Login and Register Screens
